Pulls Mac Address Range plugin support out
Refactors the quark plugin by moving Mac Address Ranges and associated
tests into separate submodules.

Also includes a hack for a pretty heinous bug. Check out
https://github.com/jkoelker/quark/issues/130 for details.
