+++ quark/tests/test_ipam.py
@@ -22,6 +22,7 @@ from oslo.config import cfg
 
 from quark.db import models
 import quark.ipam
+import quark.plugin
 
 from quark.tests import test_base
 
@@ -34,6 +35,10 @@ class QuarkIpamBaseTest(test_base.TestBase):
         neutron_db_api.configure_db()
         models.BASEV2.metadata.create_all(neutron_session._ENGINE)
         self.ipam = quark.ipam.QuarkIpam()
+
+        # FIXME(mdietz): refactor around issue #130 and remove this
+        # Ensures that the perhaps_generate_uuid event handler is initialized
+        self.plugin = quark.plugin.Plugin()
 
     def tearDown(self):
         neutron_db_api.clear_db()
@@ -348,7 +353,7 @@ class QuarkIPAddressAllocateDeallocated(QuarkIpamBaseTest):
             False, subnet, address, addresses_found
         ) as (choose_subnet):
             ipaddress = self.ipam.allocate_ip_address(self.context, 0, 0, 0)
-            self.assertIsNone(ipaddress['id'])
+            self.assertIsNotNone(ipaddress['id'])
             self.assertTrue(choose_subnet.called)
 
     def test_allocate_finds_gap_in_address_space(self):
@@ -368,7 +373,7 @@ class QuarkIPAddressAllocateDeallocated(QuarkIpamBaseTest):
         ) as (choose_subnet):
             ipaddress = self.ipam.allocate_ip_address(self.context, 0, 0, 0)
             self.assertEqual(ipaddress["address"], 2)
-            self.assertIsNone(ipaddress['id'])
+            self.assertIsNotNone(ipaddress['id'])
             self.assertTrue(choose_subnet.called)
