+++ quark/tests/test_quark_plugin.py
@@ -1568,128 +1568,6 @@ class TestQuarkDisassociatePort(TestQuarkPlugin):
                 self.plugin.disassociate_port(self.context, 1, 1)
 
 
-class TestQuarkGetMacAddressRanges(TestQuarkPlugin):
-    @contextlib.contextmanager
-    def _stubs(self, mac_range):
-        db_mod = "quark.db.api"
-        with mock.patch("%s.mac_address_range_find" % db_mod) as mar_find:
-            mar_find.return_value = mac_range
-            yield
-
-    def test_find_mac_ranges(self):
-        mar = dict(id=1, cidr="AA:BB:CC/24")
-        with self._stubs([mar]):
-            res = self.plugin.get_mac_address_ranges(self.context)
-            self.assertEqual(res[0]["id"], mar["id"])
-            self.assertEqual(res[0]["cidr"], mar["cidr"])
-
-    def test_find_mac_range(self):
-        mar = dict(id=1, cidr="AA:BB:CC/24")
-        with self._stubs(mar):
-            res = self.plugin.get_mac_address_range(self.context, 1)
-            self.assertEqual(res["id"], mar["id"])
-            self.assertEqual(res["cidr"], mar["cidr"])
-
-    def test_find_mac_range_fail(self):
-        with self._stubs(None):
-            with self.assertRaises(quark_exceptions.MacAddressRangeNotFound):
-                self.plugin.get_mac_address_range(self.context, 1)
-
-
-class TestQuarkCreateMacAddressRanges(TestQuarkPlugin):
-    @contextlib.contextmanager
-    def _stubs(self, mac_range):
-        db_mod = "quark.db.api"
-        with mock.patch("%s.mac_address_range_create" % db_mod) as mar_create:
-            mar_create.return_value = mac_range
-            yield
-
-    def test_create_range(self):
-        mar = dict(mac_address_range=dict(id=1, cidr="AA:BB:CC/24"))
-        with self._stubs(mar["mac_address_range"]):
-            res = self.plugin.create_mac_address_range(self.context, mar)
-            self.assertEqual(res["id"], mar["mac_address_range"]["id"])
-            self.assertEqual(res["cidr"], mar["mac_address_range"]["cidr"])
-
-    def test_to_mac_range_cidr_format(self):
-        cidr, first, last = self.plugin._to_mac_range("AA:BB:CC/24")
-        first_mac = str(netaddr.EUI(first, dialect=netaddr.mac_unix))
-        last_mac = str(netaddr.EUI(last, dialect=netaddr.mac_unix))
-        self.assertEqual(cidr, "AA:BB:CC:00:00:00/24")
-        self.assertEqual(first_mac, "aa:bb:cc:0:0:0")
-        self.assertEqual(last_mac, "aa:bb:cd:0:0:0")
-
-    def test_to_mac_range_just_prefix(self):
-        cidr, first, last = self.plugin._to_mac_range("AA:BB:CC")
-        first_mac = str(netaddr.EUI(first, dialect=netaddr.mac_unix))
-        last_mac = str(netaddr.EUI(last, dialect=netaddr.mac_unix))
-        self.assertEqual(cidr, "AA:BB:CC:00:00:00/24")
-        self.assertEqual(first_mac, "aa:bb:cc:0:0:0")
-        self.assertEqual(last_mac, "aa:bb:cd:0:0:0")
-
-    def test_to_mac_range_unix_format(self):
-        cidr, first, last = self.plugin._to_mac_range("AA-BB-CC")
-        first_mac = str(netaddr.EUI(first, dialect=netaddr.mac_unix))
-        last_mac = str(netaddr.EUI(last, dialect=netaddr.mac_unix))
-        self.assertEqual(cidr, "AA:BB:CC:00:00:00/24")
-        self.assertEqual(first_mac, "aa:bb:cc:0:0:0")
-        self.assertEqual(last_mac, "aa:bb:cd:0:0:0")
-
-    def test_to_mac_range_unix_cidr_format(self):
-        cidr, first, last = self.plugin._to_mac_range("AA-BB-CC/24")
-        first_mac = str(netaddr.EUI(first, dialect=netaddr.mac_unix))
-        last_mac = str(netaddr.EUI(last, dialect=netaddr.mac_unix))
-        self.assertEqual(cidr, "AA:BB:CC:00:00:00/24")
-        self.assertEqual(first_mac, "aa:bb:cc:0:0:0")
-        self.assertEqual(last_mac, "aa:bb:cd:0:0:0")
-
-    def test_to_mac_prefix_too_short_fails(self):
-        with self.assertRaises(quark_exceptions.InvalidMacAddressRange):
-            cidr, first, last = self.plugin._to_mac_range("AA-BB")
-
-    def test_to_mac_prefix_too_long_fails(self):
-        with self.assertRaises(quark_exceptions.InvalidMacAddressRange):
-            cidr, first, last = self.plugin._to_mac_range("AA-BB-CC-DD-EE-F0")
-
-    def test_to_mac_prefix_is_garbage_fails(self):
-        with self.assertRaises(quark_exceptions.InvalidMacAddressRange):
-            cidr, first, last = self.plugin._to_mac_range("F0-0-BAR")
-
-
-class TestQuarkDeleteMacAddressRanges(TestQuarkPlugin):
-    @contextlib.contextmanager
-    def _stubs(self, mac_range):
-        db_mod = "quark.db.api"
-        with contextlib.nested(
-            mock.patch("%s.mac_address_range_find" % db_mod),
-            mock.patch("%s.mac_address_range_delete" % db_mod),
-        ) as (mar_find, mar_delete):
-            mar_find.return_value = mac_range
-            yield mar_delete
-
-    def test_mac_address_range_delete_not_found(self):
-        with self._stubs(None):
-            with self.assertRaises(quark_exceptions.MacAddressRangeNotFound):
-                self.plugin.delete_mac_address_range(self.context, 1)
-
-    def test_mac_address_range_delete_in_use(self):
-        mar = mock.MagicMock()
-        mar.id = 1
-        mar.allocated_macs = 1
-        with self._stubs(mar):
-            with self.assertRaises(quark_exceptions.MacAddressRangeInUse):
-                self.plugin.delete_mac_address_range(self.context, 1)
-
-    def test_mac_address_range_delete_success(self):
-        mar = mock.MagicMock()
-        mar.id = 1
-        mar.allocated_macs = 0
-        with self._stubs(mar) as mar_delete:
-            resp = self.plugin.delete_mac_address_range(self.context, 1)
-            self.assertIsNone(resp)
-            mar_delete.assert_called_once_with(self.context, mar)
-
-
 class TestQuarkGetRoutes(TestQuarkPlugin):
     @contextlib.contextmanager
     def _stubs(self, routes):
@@ -1891,7 +1769,6 @@ class TestQuarkCreateSecurityGroup(TestQuarkPlugin):
             result = self.plugin.create_security_group(
                 self.context, {'security_group': group})
             self.assertTrue(group_create.called)
-            print "expected: %s but got: %s" % (expected, result)
             for key in expected.keys():
                 self.assertEqual(result[key], expected[key])
 
@@ -2004,7 +1881,6 @@ class TestQuarkCreateSecurityGroupRule(TestQuarkPlugin):
             result = self.plugin.create_security_group_rule(
                 self.context, {'security_group_rule': rule})
             self.assertTrue(rule_create.called)
-            print "expected: %s but got: %s" % (expected, result)
             for key in expected.keys():
                 self.assertEqual(expected[key], result[key])
 
@@ -2072,7 +1948,6 @@ class TestQuarkDeleteSecurityGroupRule(TestQuarkPlugin):
         if rule:
             dbrule = models.SecurityGroupRule()
             dbrule.update(dict(rule, group=dbgroup))
-        print rule
 
         with contextlib.nested(
                 mock.patch("quark.db.api.security_group_find"),
